+++ dwds/lib/src/debugging/location.dart
@@ -3,7 +3,6 @@
 // BSD-style license that can be found in the LICENSE file.
 
 import 'package:async/async.dart';
-import 'package:dwds/src/loaders/require.dart';
 import 'package:logging/logging.dart';
 import 'package:path/path.dart' as p;
 import 'package:source_maps/parser.dart';
@@ -162,18 +161,13 @@ class Locations {
 
   /// Returns all [Location] data for a provided JS server path.
   Future<Set<Location>> locationsForUrl(String url) async {
-    //final uri = Uri.parse(url);
-    //final segments = uri.pathSegments.first == '/'? uri.pathSegments.skip(1): uri.pathSegments;
-    //final serverPath = p.joinAll(segments);
-    
+    if (url.isEmpty) return {};
+
     final dartUri = DartUri(url, _root);
     final serverPath = dartUri.serverPath;
-    _logger.info('XXX looking for locations for js $url(server path $serverPath)');
-    
-    final module = await globalLoadStrategy.moduleForServerPath(
-        _entrypoint, serverPath);
+    final module =
+        await globalLoadStrategy.moduleForServerPath(_entrypoint, serverPath);
 
-    _logger.info('XXX module for url: $url($serverPath): $module');
     final cache = _moduleToLocations[module];
     if (cache != null) return cache;
     if (module != null) {
@@ -187,10 +181,7 @@ class Locations {
   /// The [line] number is 1-based.
   Future<Location?> locationForDart(DartUri uri, int line, int column) async {
     final locations = await locationsForDart(uri.serverPath);
-    _logger.fine('XXX locations for dart $uri:$line:$column: $locations');
-    final ret = _bestDartLocation(locations, line, column);
-    _logger.info('XXX Best location for dart ${uri.serverPath}:$line:$column: $ret');
-    return ret;
+    return _bestDartLocation(locations, line, column);
   }
 
   /// Find the [Location] for the given JS source position.
@@ -198,10 +189,7 @@ class Locations {
   /// The [line] number is 0-based.
   Future<Location?> locationForJs(String url, int line, int? column) async {
     final locations = await locationsForUrl(url);
-    _logger.fine('XXX locations for js $url:$line:$column: $locations');
-    final ret = _bestJsLocation(locations, line, column);
-    _logger.info('XXX Best location for js $url:$line:$column: $ret');
-    return ret;
+    return _bestJsLocation(locations, line, column);
   }
 
   /// Find closest existing Dart location for the line and column.
@@ -302,11 +290,8 @@ class Locations {
           await globalLoadStrategy.sourceMapPathForModule(_entrypoint, module);
       final sourceMapContents =
           await _assetReader.sourceMapContents(sourceMapPath);
-      final scriptLocation = p.url.dirname('/${relativizePath(modulePath)}');
-
-      _logger.info('XXX modulePath: $modulePath');
-      _logger.info('XXX sourceMapPath: $sourceMapPath');
-      _logger.info('XXX sourceMapContents: ${sourceMapContents != null}');
+      final scriptLocation =
+          p.url.dirname('/${stripLeadingSlashes(modulePath)}');
 
       if (sourceMapContents == null) return result;
       // This happens to be a [SingleMapping] today in DDC.
@@ -322,26 +307,10 @@ class Locations {
             // This works on Windows because path treats both / and \ as separators.
             // It will fail if the path has both separators in it.
             final relativeSegments = p.split(mapping.urls[index]);
-            
             final path = p.url.normalize(
                 p.url.joinAll([scriptLocation, ...relativeSegments]));
 
-            // detect reference to another package:
-            // scriptLocation: '/packages/home/pages'
-            // relativeSegments: '../../../account_presentation/lib/controller/sign_in_controller.dart
-            //if (scriptLocation.startsWith('/packages/') && !path.startsWith('/packages/')) {
-            //  path = "/packages${path.replaceAll('/lib/', '/')}";
-            //}
-
-            //if (scriptLocation.startsWith('/packages/') && relativeSegments.contains('lib') && relativeSegments.first == '..') {
-            //  relativeSegments = relativeSegments.skip(1).where((e) => e != 'lib').toList();
-            //}
             final dartUri = DartUri(path, _root);
-            _logger.info('XXX adding location for $path($dartUri)');
-            //if (path.contains('sign_in_controller')) {
-             // _logger.info('XXX script location: $scriptLocation, relative segments: $relativeSegments');
-            // _logger.info('XXX add location $modulePath:$lineEntry:$entry at $path($dartUri)');
-            //}
             result.add(Location.from(
               modulePath,
               lineEntry,
@@ -356,9 +325,6 @@ class Locations {
             .putIfAbsent(
                 location.dartLocation.uri.serverPath, () => <Location>{})
             .add(location);
-      }
-      if (module.contains('welcome_page.dart')) {
-        _logger.info('XXX locations for module $module computed: $result');
       }
       return _moduleToLocations[module] = result;
     });