+++ dwds/lib/src/debugging/metadata/provider.dart
@@ -6,7 +6,6 @@ import 'dart:async';
 import 'dart:convert';
 
 import 'package:async/async.dart';
-import 'package:dwds/src/utilities/dart_uri.dart';
 import 'package:logging/logging.dart';
 import 'package:path/path.dart' as p;
 
@@ -220,12 +219,8 @@ class MetadataProvider {
   }
 
   void _addMetadata(ModuleMetadata metadata) {
-    //final modulePath = DartUri(metadata.moduleUri).serverPath;
-    //final sourceMapPath = DartUri(metadata.sourceMapUri).serverPath;
     final modulePath = stripLeadingSlashes(metadata.moduleUri);
     final sourceMapPath = stripLeadingSlashes(metadata.sourceMapUri);
-
-    _logger.info('XXX Adding module: ${metadata.name} with url: ${metadata.moduleUri} and server path $modulePath');
 
     _moduleToSourceMap[metadata.name] = sourceMapPath;
     _modulePathToModule[modulePath] = metadata.name;
@@ -236,9 +231,6 @@ class MetadataProvider {
         throw AbsoluteImportUriException(library.importUri);
       }
       _libraries.add(library.importUri);
-      if (library.importUri.contains('sign_in_controller')) {
-        _logger.info('XXX loading library info for ${library.importUri}');
-      }
       _scripts[library.importUri] = [];
 
       _scriptToModule[library.importUri] = metadata.name;