+++ dwds/lib/src/loaders/frontend_server_require.dart
@@ -2,7 +2,6 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-import 'package:logging/logging.dart';
 import 'package:path/path.dart' as p;
 
 import '../debugging/metadata/provider.dart';
@@ -13,7 +12,6 @@ import 'require.dart';
 
 /// Provides a [RequireStrategy] suitable for use with Frontend Server.
 class FrontendServerRequireStrategyProvider {
-  final _logger = Logger('FrontendServerRequireStrategyProvider');
   final ReloadConfiguration _configuration;
   final AssetReader _assetReader;
   final PackageUriMapper _packageUriMapper;
@@ -32,40 +30,36 @@ class FrontendServerRequireStrategyProvider {
     _assetReader,
   );
 
-  FrontendServerRequireStrategyProvider(this._configuration, this._assetReader, this._packageUriMapper,
-      this._digestsProvider, this._basePath,);
+  FrontendServerRequireStrategyProvider(
+    this._configuration,
+    this._assetReader,
+    this._packageUriMapper,
+    this._digestsProvider,
+    this._basePath,
+  );
 
   RequireStrategy get strategy => _requireStrategy;
 
   String _removeBasePath(String path) {
     if (_basePath.isEmpty) return path;
-    // If path is a server path it might start with a '/'.
-    // final base = path.startsWith('/') ? '/$_basePath/' : _basePath;
 
-    final stripped = relativizePath(path);
-    return relativizePath(stripped.substring(_basePath.length));
-    //return path.startsWith(base) ? path.substring(base.length) : path;
+    final stripped = stripLeadingSlashes(path);
+    return stripLeadingSlashes(stripped.substring(_basePath.length));
   }
 
   String _addBasePath(String serverPath) => _basePath.isEmpty
-      ? relativizePath(serverPath)
-      : '$_basePath/${relativizePath(serverPath)}';
+      ? stripLeadingSlashes(serverPath)
+      : '$_basePath/${stripLeadingSlashes(serverPath)}';
 
   Future<Map<String, String>> _moduleProvider(
           MetadataProvider metadataProvider) async =>
       (await metadataProvider.moduleToModulePath).map((key, value) =>
-          MapEntry(key, relativizePath(removeJsExtension(value))));
+          MapEntry(key, stripLeadingSlashes(removeJsExtension(value))));
 
   Future<String?> _moduleForServerPath(
       MetadataProvider metadataProvider, String serverPath) async {
     final modulePathToModule = await metadataProvider.modulePathToModule;
     final relativeServerPath = _removeBasePath(serverPath);
-    modulePathToModule.forEach((key, value) {
-      if (key.contains(relativeServerPath) || serverPath.contains(key)) {
-      _logger.info('XXX inexact match: serverPath: $serverPath($relativeServerPath), key: $key');
-      }
-    });
-    _logger.info('XXX exact match: serverPath: $serverPath, mpd: ${modulePathToModule[relativeServerPath]}');
     return modulePathToModule[relativeServerPath];
   }
 
@@ -83,10 +77,10 @@ class FrontendServerRequireStrategyProvider {
       return _addBasePath(appUri.path);
     }
     if (appUri.isScheme('package')) {
-      //final resolved = _assetReader.resolvePackageUrl(appUri);
       final resolved = _packageUriMapper.packageUriToServerPath(appUri);
-      //_logger.info('XXX: server path for $appUrl: $resolved');
-      return resolved;
+      if (resolved != null) {
+        return _addBasePath(resolved);
+      }
     }
     return null;
   }
@@ -106,4 +100,3 @@ class FrontendServerRequireStrategyProvider {
     return result;
   }
 }
-