+++ dwds/lib/src/readers/proxy_server_asset_reader.dart
@@ -70,8 +70,5 @@ class ProxyServerAssetReader implements AssetReader {
       _readResource(serverPath);
 
   @override
-  String? resolvePackageUrl(Uri packageUri) => null;
-
-  @override
   Future<void> close() async => _client.close();
 }