+++ fixtures/_testCircular1/lib/library1.dart
@@ -1,11 +1,13 @@
-// Copyright (c) 2021, the Dart project authors.  Please see the AUTHORS file
+// Copyright (c) 2019, the Dart project authors.  Please see the AUTHORS file
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-/// A library that we can import.
-library test_library;
+// @dart=2.9
 
-int aVariable = 3;
+/// A library that we can import.
+library _test_circular1;
+
+import 'package:_test_circular2/library2.dart';
 
 String concatenate(String a, String b) {
   return '$a$b'; // Breakpoint: Concatenate